# tui/api/v3/admin/admin.openapi.yaml
openapi: 3.0.3
info:
  title: Agent Harbor Admin API
  version: 0.1.0
  description: |
    Local administrator contract for configuration, readiness, sessions,
    runtime status, provider-secret operations, target control, and shutdown.
    Unix-socket peer authorization occurs before HTTP dispatch. An explicitly
    enabled loopback TCP listener requires the AdminToken security scheme.
  license:
    name: MIT
    url: https://opensource.org/license/mit
...
servers:
  - url: http://localhost
    description: Unix socket or explicitly enabled loopback Admin listener
...
paths:
  "/v1/readiness":
    get:
      operationId: getReadiness
      tags:
        - instance
      x-requirements:
        - V3-SCP-002
        - V3-SCP-003
        - V3-CAP-003
      summary: Get exact instance readiness
      x-error-codes:
        - operation_error
      responses:
        '200':
          description: Exact ready instance identity
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Readiness"
        default:
          $ref: "#/components/responses/Phase04Problem"
        "500":
          $ref: "#/components/responses/Phase04Problem"
  "/v1/status":
    get:
      operationId: getRuntimeStatus
      tags:
        - instance
      x-requirements:
        - V3-CAP-011
        - V3-OBS-001
        - V3-API-005
      summary: Get masked runtime status
      x-error-codes:
        - operation_error
      responses:
        '200':
          description: Masked complete runtime status snapshot
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/RuntimeStatus"
        default:
          $ref: "#/components/responses/Phase04Problem"
        "500":
          $ref: "#/components/responses/Phase04Problem"
  "/v1/events":
    get:
      operationId: streamRuntimeEvents
      tags:
        - instance
      x-requirements:
        - V3-CAP-011
        - V3-CAP-021
        - V3-API-005
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
            pattern: "^[0-9]+$"
      summary: Stream runtime observation events
      x-observation-cursor:
        id_format: unsigned_decimal
        stream:
          last_event_id_is_exclusive: true
          emitted_ids_are_strictly_monotonic: true
          stale_cursor_event: resync_required
          stale_cursor_closes_stream: true
          client_cursor_never_regresses: true
        cross_endpoint_snapshot:
          runtime_status_field: RuntimeStatus.observations.last_event_id
          page_field: ObservationPage.next_cursor
          frozen_until: required_listObservations_until_query
          resume_cursor: terminal_committed_observation_page_cursor_equal_to_until
          reason: concurrent_events_above_until_are_replayed_by_the_exclusive_stream
      x-error-codes:
        - validation_error
        - operation_error
      responses:
        '200':
          description: |
            Server-sent events. Each event has decimal `id`, an event name equal
            to RuntimeEvent.type, and one JSON RuntimeEvent object in `data`.
            A stale cursor returns event `resync_required` then closes.
          content:
            text/event-stream:
              schema:
                "$ref": "#/components/schemas/RuntimeEvent"
        default:
          $ref: "#/components/responses/Phase04Problem"
        "500":
          $ref: "#/components/responses/Phase04Problem"
...
  "/v1/config":
    get:
      operationId: getConfig
      responses:
        "200":
          description: Complete mutable configuration view with masked credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutableConfigSnapshot"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: patchConfig
      x-audit: config_publish
      x-cancellation: Cancellation before the publication transaction commits leaves
        the active snapshot unchanged; after commit the result is
        operation_unknown and the caller reloads readiness and config without
        replaying publication.
      x-error-codes:
        - instance_mismatch
        - generation_conflict
        - no_change
        - duplicate_id
        - validation_error
        - stage_not_found
        - stage_expired
        - stage_reserved
        - secret_prepare_failed
        - secret_cleanup_pending
        - publish_conflict
        - operation_error
        - operation_unknown
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchConfigCommand"
      responses:
        "200":
          description: Newly active immutable snapshot identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotIdentity"
        "409":
          $ref: "#/components/responses/Problem"
        "410":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
  "/v1/sessions":
    get:
      operationId: listAgentSessions
      tags:
        - sessions
      x-requirements:
        - V3-CAP-011
        - V3-SES-001
      summary: List managed agent sessions
      parameters:
        - name: instance_id
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/InstanceID"
      x-error-codes:
        - instance_mismatch
        - validation_error
        - operation_error
      responses:
        '200':
          description: AgentSessions ordered by updated_at descending then ID
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/AgentSession"
        default:
          $ref: "#/components/responses/Phase04Problem"
        "500":
          $ref: "#/components/responses/Phase04Problem"
    post:
      operationId: createAgentSession
      tags:
        - sessions
      x-requirements:
        - V3-CAP-013
        - V3-SES-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentSessionCommand"
      summary: Create a managed agent session
      x-audit: session_create
      x-cancellation: Cancellation before the create transaction commits has no side
        effect; after commit the caller reloads the exact instance session list.
      x-error-codes:
        - instance_mismatch
        - stale_precondition
        - validation_error
        - operation_error
        - operation_unknown
      responses:
        '201':
          description: Persisted AgentSession before terminal launch
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AgentSession"
        "422":
          $ref: "#/components/responses/Phase04Problem"
        default:
          $ref: "#/components/responses/Phase04Problem"
        "500":
          $ref: "#/components/responses/Phase04Problem"
...
components:
  securitySchemes:
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: opaque local development token
      description: One-time-delivered token whose SHA-256 hash is stored; required
        only on an explicitly enabled loopback TCP Admin listener
    UnixPeer:
      type: apiKey
      in: header
      name: X-Agent-Harbor-Unix-Peer-Authenticated
      description: Logical scheme satisfied from Unix-socket peer credentials before
        HTTP dispatch. The listener strips and ignores any client-supplied
        header with this name.
...
  responses:
    Problem:
      description: Typed Agent Harbor problem
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Phase04Problem:
      description: Phase 04 typed RFC 9457 problem response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Phase04Problem"
  schemas:
...
    Problem:
      "$ref": "./admin/common.schemas.yaml#/components/schemas/Problem"
...
    Readiness:
      "$ref": "./admin/runtime.schemas.yaml#/components/schemas/Readiness"
    SnapshotIdentity:
      "$ref": "./admin/common.schemas.yaml#/components/schemas/SnapshotIdentity"
    RuntimeStatus:
      "$ref": "./admin/runtime.schemas.yaml#/components/schemas/RuntimeStatus"
...
    MutableConfigSnapshot:
      $ref: "./admin/config-core.schemas.yaml#/components/schemas/MutableConfigSnapshot"
    PatchConfigCommand:
      $ref: "./admin/config-core.schemas.yaml#/components/schemas/PatchConfigCommand"
...
    AgentSession:
      "$ref": "./admin/session.schemas.yaml#/components/schemas/AgentSession"
...
    RuntimeEvent:
      "$ref": "./admin/runtime.schemas.yaml#/components/schemas/RuntimeEvent"
    InstanceID:
      $ref: ./admin/deck.schemas.yaml#/components/schemas/InstanceID
...
    CreateAgentSessionCommand:
      $ref: ./admin/deck.schemas.yaml#/components/schemas/CreateAgentSessionCommand
...
    Phase04ProblemCode:
      type: string
      enum:
        - instance_mismatch
        - stale_precondition
        - validation_error
        - invalid_lifecycle
        - tmux_identity_mismatch
        - operation_timeout
        - operation_unknown
        - resume_unavailable
        - session_process_active
        - operation_error
        - credential_not_found
        - session_not_found
        - secret_denied
        - publish_conflict
        - secret_unavailable
        - admission_rejected
        - target_not_found
        - preview_unavailable
        - attachment_unavailable
        - hook_unavailable
        - observation_cursor_stale
        - active_config_invalid
        - native_config_root_denied
        - native_config_root_conflict
        - native_hook_entry_conflict
        - native_config_write_failed
        - runtime_shutting_down
        - admin_authentication_required
        - too_many_headers
        - headers_too_large
        - invalid_request
        - not_found
        - method_not_allowed
        - unsupported_content_encoding
        - unsupported_content_type
        - body_too_large
        - invalid_json
        - missing_field
        - null_field
        - unknown_field
        - duplicate_field
        - invalid_field_type
        - invalid_value
        - unix_peer_required
    Phase04Problem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            code:
              $ref: "#/components/schemas/Phase04ProblemCode"
security:
  - UnixPeer: []
  - AdminToken: []
...